Extract search tools into tools/search.py
Move local_search (+ LocalSearchError) and the five remote search tools
(leansearch, loogle, leanfinder, state_search, hammer_premise) out of
server.py. Add server.__all__ to keep test-patched imports (e.g.
lean_local_search) from being pruned now that only tool modules use them.
